requirements.txt
scikit-learn==1.8.0
lightgbm==4.6.0
imbalanced-learn==0.14.1
numpy==2.4.4
pandas==3.0.2
joblib==1.5.3
threadpoolctl==3.6.0
sklearn-compat==0.1.5
tzdata==2026.1
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
joblib==1.5.3
